embedding_requirements.txt
supabase>=2.0.0
google-generativeai>=0.8.0
